+++ source/Cosmos.System/FileSystem/FAT/FatStream.cs
@@ -186,7 +186,9 @@ namespace Cosmos.System.FileSystem.FAT
                     xReadSize = xCount;
                 }
 
-                Array.Copy(xCluster, xPosInCluster, aBuffer, xOffset, xReadSize);
+                //TODO: .Net Core
+                //Array.Copy(xCluster, xPosInCluster, aBuffer, xOffset, xReadSize);
+                Array.Copy(xCluster, (int) xPosInCluster, aBuffer, (int) xOffset, (int) xReadSize);
 
                 xOffset += xReadSize;
                 xCount -= xReadSize;
@@ -247,7 +249,9 @@ namespace Cosmos.System.FileSystem.FAT
 
                 byte[] xCluster;
                 mFS.Read(xClusterIdx, out xCluster);
-                Array.Copy(aBuffer, aOffset, xCluster, xPosInCluster, xWriteSize);
+                //TODO: .Net Core
+                //Array.Copy(aBuffer, aOffset, xCluster, xPosInCluster, xWriteSize);
+                Array.Copy(aBuffer, aOffset, xCluster, (int) xPosInCluster, (int) xWriteSize);
                 mFS.Write(mFatTable[xClusterIdx], xCluster);
                 xOffset += xWriteSize;
                 xCount -= xWriteSize;
